requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.7
pydantic==1.10.12
python-dotenv==1.0.1
requests==2.31.0
ariadne==0.20.1
PyJWT==2.8.0
